+++ src/components/pages/dispenser/index.tsx
@@ -23,7 +23,7 @@ import {
   AsideWidgetButton,
   MainContent
 } from './styled-components'
-import { Statistics, RedirectWidget } from './components'
+import { Statistics, RedirectWidget, WhitelistWidget } from './components'
 import { TextLink } from 'components/common'
 import {
   defineDispenserStatus,
@@ -32,7 +32,13 @@ import {
   alertError
 } from 'helpers'
 import { Redirect, useHistory, useParams } from 'react-router-dom'
-import { TDispenser, TDispenserStatus, TLinkDecrypted } from 'types'
+import {
+  TDispenser,
+  TDispenserStatus,
+  TLinkDecrypted,
+  TDispenserWhitelistType,
+  TDispenserWhitelistItemAddress
+} from 'types'
 import { connect } from 'react-redux'
 import * as asyncDispensersActions from 'data/store/reducers/dispensers/async-actions'
 import { decrypt, encrypt } from 'lib/crypto'
@@ -156,6 +162,17 @@ const mapDispatcherToProps = (dispatch: IAppDispatch) => {
       successCallback,
       errorCallback
     })),
+    toggleWhitelist: (
+      dispenser_id: string,
+      whitelist_on: boolean,
+      successCallback: () => void,
+      errorCallback: () => void
+    ) => dispatch(asyncDispensersActions.toggleWhitelistOn({
+      dispenser_id,
+      whitelist_on,
+      successCallback,
+      errorCallback
+    })),
     downloadReport: (
       dispenser_id: string,
     ) => dispatch(asyncDispensersActions.downloadReport(
@@ -164,6 +181,7 @@ const mapDispatcherToProps = (dispatch: IAppDispatch) => {
   }
 }
 
+// @ts-ignore
 type ReduxType = ReturnType<typeof mapStateToProps> & ReturnType<typeof mapDispatcherToProps>
 
 const Dispenser: FC<ReduxType> = ({
@@ -180,9 +198,11 @@ const Dispenser: FC<ReduxType> = ({
   toggleRedirectURL,
   getDispenserStats,
   downloadReport,
+  toggleWhitelist,
   chainId
 }) => {
   const { id } = useParams<{id: string}>()
+  // @ts-ignore
   const dispenser: TDispenser | undefined = dispensers.find(dispenser => String(dispenser.dispenser_id) === id)
   const history = useHistory()
   const [
@@ -263,7 +283,10 @@ const Dispenser: FC<ReduxType> = ({
     multiscan_qr_id,
     title,
     links_claimed,
-    links_assigned
+    links_assigned,
+    whitelist_on,
+    whitelist_type,
+    whitelist
   } = dispenser
 
   const currentStatus = defineDispenserStatus(
@@ -360,6 +383,24 @@ const Dispenser: FC<ReduxType> = ({
           toggleRedirectURL(
             dispenser.dispenser_id as string,
             redirectOn,
+            successCallback,
+            errorCallback
+          )
+        }}
+      />
+      <WhitelistWidget
+        loading={loading}
+        isWhitelisted={whitelist_on}
+        whitelistType={whitelist_type}
+        dispenserId={dispenser_id as string}
+        toggleWhitelistOn={(
+          whitelistOn,
+          successCallback,
+          errorCallback
+        ) => {
+          toggleWhitelist(
+            dispenser_id as string,
+            whitelistOn,
             successCallback,
             errorCallback
           )