+++ src/data/store/reducers/dispensers/async-actions/index.tsx
@@ -7,12 +7,20 @@ import toggleRedirectOn from './toggle-redirect-on'
 import updateRedirectURL from './update-redirect-url'
 import getDispenserStats from './get-dispenser-stats'
 import downloadReport from './download-report'
+import toggleWhitelistOn from './toggle-whitelist-on'
+import createWhitelist from './create-whitelist'
+import updateWhitelist from './update-whitelist'
+import getDispenserWhitelist from './get-dispenser-whitelist'
 
 export {
   addDispenser,
+  updateWhitelist,
+  toggleWhitelistOn,
+  getDispenserWhitelist,
   downloadDispenserQR,
   downloadReport,
   addLinksToQR,
+  createWhitelist,
   updateDispenser,
   updateStatus,
   toggleRedirectOn,