+++ src/types/wallet.tsx
@@ -5,6 +5,7 @@ type TWallet = {
   name: string
   chains: string[]
   token_types: (TTokenType | null)[]
+  available_for_not_sponsored: boolean
 }
 
 export default TWallet